Dealt with special characters.
Changed the code to add the texts as "html" instead of just "text". This change should allow the escaped values to have the correct look and feel. The code to download the .csv results file is now using the textual data from the data set instead of taking it from the DOM nodes. Added a solution to decode/unescape the special characters that have been encoded with htmlspecialchars. Also added insturctions for running linter to the README file.